+++ src/Veneer/Proxy.php
@@ -8,7 +8,9 @@ namespace DecodeLabs\Veneer;
 
 use Psr\Container\ContainerInterface;
 
-interface Facade
+interface Proxy
 {
+    public static function setVeneerProxyTargetInstance(object $instance): void;
+    public static function getVeneerProxyTargetInstance(): ?object;
     public static function __callStatic(string $name, array $args);
 }